+++ das_client/app/integration_test/app_test_ios_3.dart
@@ -6,6 +6,7 @@ import 'package:logging/logging.dart';
 
 import 'test/automatic_advancement_test.dart' as automatic_advancement_tests;
 import 'test/navigation_test.dart' as navigation_tests;
+import 'test/train_journey_table_adl_test.dart' as train_journey_table_adl_tests;
 import 'test/train_search_test.dart' as train_search_tests;
 
 AppLocalizations l10n = AppLocalizationsDe();
@@ -16,6 +17,7 @@ void main() {
   Logger.root.level = Level.FINE;
   Logger.root.onRecord.listen(LogPrinter(appName: 'DAS IntegrationTests').call);
 
+  train_journey_table_adl_tests.main();
   navigation_tests.main();
   train_search_tests.main();
   automatic_advancement_tests.main();