+++ das_client/app/integration_test/util/test_time_constants.dart
@@ -15,4 +15,10 @@ class TestTimeConstants extends TimeConstants {
 
   @override
   int get arrivalDepartureOperationalResetSeconds => 2;
+
+  @override
+  int get adlEndDisplaySeconds => 2;
+
+  @override
+  int get kmDecisiveGradientResetSeconds => 2;
 }