+++ das_client/app/lib/pages/journey/train_journey_view_model.dart
@@ -1,8 +1,10 @@
 import 'dart:async';
 
+import 'package:app/di/di.dart';
 import 'package:app/pages/journey/train_journey/automatic_advancement_controller.dart';
 import 'package:app/pages/journey/train_journey/widgets/table/config/train_journey_settings.dart';
 import 'package:app/util/error_code.dart';
+import 'package:app/util/time_constants.dart';
 import 'package:flutter/material.dart';
 import 'package:logging/logging.dart';
 import 'package:rxdart/rxdart.dart';
@@ -21,6 +23,7 @@ class TrainJourneyViewModel {
     _init();
   }
 
+  final _resetToKmAfterSeconds = DI.get<TimeConstants>().kmDecisiveGradientResetSeconds;
   static const _warnappWindowMilliseconds = 1250;
 
   final SferaRemoteRepo _sferaRemoteRepo;
@@ -36,11 +39,18 @@ class TrainJourneyViewModel {
 
   Stream<WarnappEvent> get warnappEvents => _rxWarnapp.stream;
 
+  Stream<bool> get showDecisiveGradient => _rxShowDecisiveGradient.distinct();
+
+  bool get showDecisiveGradientValue => _rxShowDecisiveGradient.value;
+
   AutomaticAdvancementController automaticAdvancementController = AutomaticAdvancementController();
 
   final _rxSettings = BehaviorSubject<TrainJourneySettings>.seeded(TrainJourneySettings());
   final _rxErrorCode = BehaviorSubject<ErrorCode?>.seeded(null);
   final _rxWarnapp = PublishSubject<WarnappEvent>();
+
+  final _rxShowDecisiveGradient = BehaviorSubject<bool>.seeded(false);
+  Timer? _showDecisiveGradientTimer;
 
   DateTime? _lastWarnappEventTimestamp;
 
@@ -66,6 +76,7 @@ class TrainJourneyViewModel {
         case SferaRemoteRepositoryState.disconnected:
           WakelockPlus.disable();
           _disableWarnapp();
+          _resetSettings();
           if (_sferaRemoteRepo.lastError != null) {
             _rxErrorCode.add(ErrorCode.fromSfera(_sferaRemoteRepo.lastError!));
             setAutomaticAdvancement(false);
@@ -133,7 +144,10 @@ class TrainJourneyViewModel {
   void dispose() {
     _rxSettings.close();
     _rxWarnapp.close();
+    _rxShowDecisiveGradient.close();
     _stateSubscription?.cancel();
+    _warnappSignalSubscription?.cancel();
+    _warnappAbfahrtSubscription?.cancel();
     automaticAdvancementController.dispose();
   }
 
@@ -145,6 +159,18 @@ class TrainJourneyViewModel {
         now.difference(_lastWarnappEventTimestamp!).inMilliseconds < _warnappWindowMilliseconds) {
       _log.info('Abfahrt detected while warnapp message was within $_warnappWindowMilliseconds ms -> Warning!');
       _rxWarnapp.add(WarnappEvent());
+    }
+  }
+
+  void toggleKmDecisiveGradient() {
+    if (!showDecisiveGradientValue) {
+      _rxShowDecisiveGradient.add(true);
+      _showDecisiveGradientTimer = Timer(Duration(seconds: _resetToKmAfterSeconds), () {
+        if (_rxShowDecisiveGradient.value) _rxShowDecisiveGradient.add(false);
+      });
+    } else {
+      _rxShowDecisiveGradient.add(false);
+      _showDecisiveGradientTimer?.cancel();
     }
   }
 }